Message reporting: Fixed notes being reported out-of-order
The previous fix for unordered error reporting by sorting the messages by
column and line numbers broke notes, as note line/column numbers are in
general case out-of-order by default. This meant that the sorting put the
notes into wrong order, causing the errors messages to be incredibly
confusing.

Notes are now correctly attached to the error or warning they belong.

Signed-off-by: Valtis <erkka_kaaria@outlook.com>